+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.14.1...HEAD) - TBD
 
+### :bug: Fixed
+
+- `AttributeError` on `DoctestItem` when schemathesis plugin is loaded. [#3663](https://github.com/schemathesis/schemathesis/issues/3663)
+
 ## [4.14.1](https://github.com/schemathesis/schemathesis/compare/v4.14.0...v4.14.1) - 2026-03-27
 
 ### :bug: Fixed